+++ e2e/docs/.vuepress/components/ArticleList.vue
@@ -54,7 +54,6 @@ defineProps<{
         <span v-if="info.tag" class="tag">Tag: {{ info.tag.join(', ') }}</span>
       </div>
 
-      <!-- eslint-disable-next-line vue/no-v-html -->
       <div v-if="info.excerpt" class="excerpt" v-html="info.excerpt" />
     </article>
   </div>
@@ -119,15 +118,9 @@ defineProps<{
       transform: scaleX(0);
     }
 
-    &:hover {
-      &::after {
-        visibility: visible;
-        transform: scaleX(1);
-      }
-    }
-
-    a {
-      color: inherit;
+    &:hover::after {
+      visibility: visible;
+      transform: scaleX(1);
     }
   }