+++ plugins/blog/plugin-comment/src/client/components/WalineComment.ts
@@ -1,3 +1,4 @@
+import type { ExactLocaleConfig } from '@vuepress/helper/client'
 import { LoadingIcon, useLocaleConfig, wait } from '@vuepress/helper/client'
 import { watchImmediate } from '@vueuse/core'
 import { pageviewCount } from '@waline/client/pageview'
@@ -10,10 +11,10 @@ import {
   nextTick,
   onMounted,
 } from 'vue'
-import { usePageFrontmatter, usePageLang } from 'vuepress/client'
+import { ClientOnly, usePageFrontmatter, usePageLang } from 'vuepress/client'
 import type {
   CommentPluginFrontmatter,
-  WalineLocaleConfig,
+  WalineLocaleData,
 } from '../../shared/index.js'
 import { useWalineOptions } from '../helpers/index.js'
 
@@ -21,7 +22,7 @@ import '@waline/client/waline.css'
 import '../styles/waline.css'
 
 declare const WALINE_META: boolean
-declare const WALINE_LOCALES: WalineLocaleConfig
+declare const WALINE_LOCALES: ExactLocaleConfig<WalineLocaleData>
 
 const walineLocales = WALINE_LOCALES
 
@@ -103,15 +104,15 @@ export default defineComponent({
             { id: 'comment', class: 'waline-wrapper' },
             h(
               defineAsyncComponent({
-                loader: async () =>
-                  (
-                    await import(
-                      /* webpackChunkName: "waline" */ '@waline/client/component'
-                    )
-                  ).Waline,
+                loader: async () => {
+                  const { Waline } = await import(
+                    /* webpackChunkName: "waline" */ '@waline/client/component'
+                  )
+
+                  return () => h(ClientOnly, () => h(Waline, walineProps.value))
+                },
                 loadingComponent: LoadingIcon,
               }),
-              walineProps.value,
             ),
           )
         : null