+++ plugins/development/plugin-toc/src/client/components/Toc.ts
@@ -104,15 +104,9 @@ export const Toc = defineComponent({
   name: 'Toc',
 
   props: {
-    headers: {
-      type: Array as PropType<TocPropsHeaders | undefined>,
-      default: undefined,
-    },
+    headers: Array as PropType<TocPropsHeaders | undefined>,
 
-    options: {
-      type: Object as PropType<Partial<TocPropsOptions>>,
-      default: () => ({}),
-    },
+    options: Object as PropType<Partial<TocPropsOptions>>,
   },
 
   setup(props) {