+++ plugins/markdown/plugin-shiki/src/node/shikiPlugin.ts
@@ -1,3 +1,4 @@
+import { isModuleAvailable } from '@vuepress/helper'
 import type {
   MarkdownItCollapsedLinesOptions,
   MarkdownItLineNumbersOptions,
@@ -8,6 +9,7 @@ import {
 } from '@vuepress/highlighter-helper'
 import type { Plugin } from 'vuepress/core'
 import { isPlainObject } from 'vuepress/shared'
+import { colors } from 'vuepress/utils'
 import { createMarkdownFilePathGetter } from './markdown/highlighter/createMarkdownFilePathGetter.js'
 import type { MarkdownItPreWrapperOptions } from './markdown/index.js'
 import {
@@ -18,6 +20,7 @@ import {
 } from './markdown/index.js'
 import type { ShikiPluginOptions } from './options.js'
 import { prepareClientConfigFile } from './prepareClientConfigFile.js'
+import { TWOSLASH_RE, logger } from './utils.js'
 
 export const shikiPlugin = (_options: ShikiPluginOptions = {}): Plugin => {
   return (app) => {
@@ -34,18 +37,52 @@ export const shikiPlugin = (_options: ShikiPluginOptions = {}): Plugin => {
     options.lineNumbers ??= true
     options.collapsedLines ??= 'disable'
 
+    if (
+      options.twoslash &&
+      !isModuleAvailable('@vuepress/shiki-twoslash', import.meta)
+    ) {
+      logger.error(
+        `${colors.cyan('twoslash')} is enabled, but ${colors.magenta('@vuepress/shiki-twoslash')} is not installed, please install it manually`,
+      )
+      options.twoslash = false
+    }
+
+    /**
+     * Whether to enable the `v-pre` configuration of the code block
+     */
+    let enableVPre = true
+
     return {
       name: '@vuepress/plugin-shiki',
+
+      extendsMarkdownOptions: (opts) => {
+        /**
+         * Turn off the `v-pre` configuration of the code block.
+         */
+        if (opts.vPre !== false) {
+          const vPre = isPlainObject(opts.vPre) ? opts.vPre : { block: true }
+          if (vPre.block) {
+            opts.vPre ??= {}
+            opts.vPre.block = false
+          }
+          enableVPre = vPre.block ?? true
+        } else {
+          enableVPre = false
+        }
+      },
 
       extendsMarkdown: async (md) => {
         const { preWrapper, lineNumbers, collapsedLines } = options
 
         const markdownFilePathGetter = createMarkdownFilePathGetter(md)
-        const shikiHighlighter = await createShikiHighlighter(options)
+        const { highlighter, loadLang, extraTransformers } =
+          await createShikiHighlighter(app, options, enableVPre)
 
         md.options.highlight = getHighLightFunction(
-          shikiHighlighter,
+          highlighter,
           options,
+          extraTransformers,
+          loadLang,
           markdownFilePathGetter,
         )
 
@@ -54,6 +91,11 @@ export const shikiPlugin = (_options: ShikiPluginOptions = {}): Plugin => {
         if (preWrapper) {
           md.use<MarkdownItLineNumbersOptions>(lineNumbersPlugin, {
             lineNumbers,
+            resolveLineNumbers(info) {
+              return options.twoslash && TWOSLASH_RE.test(info)
+                ? false
+                : undefined
+            },
           })
           md.use<MarkdownItCollapsedLinesOptions>(collapsedLinesPlugin, {
             collapsedLines,