+++ tools/helper/src/client/composables/index.ts
@@ -1,3 +1,4 @@
-export * from './useDarkmode.js'
+export * from './useDarkMode.js'
 export * from './useLocaleConfig.js'
+export * from './useKeys.js'
 export * from './useRoutePaths.js'